Create american-latino-heritage.geojson
Since the age of exploration, Latino peoples have played a profoundly important role in American history, contributing to and shaping the rich heritage of the United States in many ways for some 500 years. Explore their stories and their legacy by visiting the units of the National Park System and places listed in the National Register of Historic Places, most of which are designated as National Historic Landmarks, throughout the nation featured in this itinerary.The National Park Service preserves historic places and stories of America’s diverse cultural heritage and expands and maintains the National Register of Historic Places. The American Latino Heritage Discover Our Shared Heritage Travel Itinerary was produced by the National Park Service’s Heritage Education Services in partnership with the National Conference of State Historic Preservation Officers.