Create amistad.geojson
Amistad - Seeking Freedom in Connecticut

The National Park Service's National Register of Historic Places in partnership with the communities of Farmington, Hartford, New Haven and New London, Connecticut, the Connecticut Commission on Culture & Tourism and the National Conference of State Historic Preservation Officers invite you to explore Amistad: Seeking Freedom in Connecticut. The plight of the Mende Africans caught the attention of America in the 1840s, and continues to hold our attention today. This travel itinerary highlights 14 historic places listed in the National Register of Historic Places that tell the story of the Amistad and the Mende Africans' legal battle and quest for freedom in Connecticut.