Create asheville.geojson
Asheville, North Carolina

The National Park Service's National Register of Historic Places in partnership with the City of Asheville, Buncombe County, Asheville Convention & Visitors Bureau, Thomas Wolfe House, North Carolina State Historic Preservation Office and the National Conference of State Historic Preservation Officers (NCSHPO), proudly invite you to explore Asheville, North Carolina. Nestled in the Blue Ridge Mountains of Western North Carolina, Asheville is perhaps best known as the location of George Vanderbilt's Biltmore Estate and the home of major American novelist, Thomas Wolfe. Asheville is also a thriving and eclectic city, dubbed the “Paris of the South” in the early 1900s for establishing itself as an artisan city with unique style and architectural talent. This travel itinerary highlights 45 historic places listed in the National Register of Historic Places that explore the development of Asheville and Buncombe County as an area of world-class resorts and mountain retreats.