Create underground-railroad.geojson
Aboard the Underground Railroad

The Underground Railroad refers to the effort--sometimes spontaneous, sometimes highly organized--to assist persons held in bondage in North America to escape from slavery. Historic places along the Underground Railroad are testament of African American capabilities. The network provided an opportunity for sympathetic white Americans to play a role in resisting slavery, and brought together, however uneasily at times, men and women of both races to begin to set aside assumptions about the other race and to work together on issues of mutual concern. At the most dramatic level, the Underground Railroad provided stories of guided escapes from the South, rescues of arrested fugitives in the North, complex communication systems, and individual acts of bravery and suffering in the quest for freedom for all.