Create atlanta.geojson
Atlanta

The National Park Service's National Register of Historic Places and Southeast Regional Office, in conjunction with the Atlanta History Center, the Historic Preservation Division of the Georgia Department of Natural Resources, and the National Conference of State Historic Preservation Officers (NCSHPO), proudly invite you to explore Atlanta, Georgia. Atlanta began as the terminal point of the Western and Atlantic Railroad, a project authorized by the State of Georgia in 1836. Originally known as Terminus, and later Marthasville, by the Civil War Atlanta was a bustling city. Crippled by the burning of the city during the war, Atlanta rebounded during the last part of the century. Today it is home to more than 4 million people and is considered the entertainment and cultural center of the South, attracting more than 17 million travelers each year.