Create augusta.geojson
History of Augusta

Founded in 1736 on the western bank of the Savannah River, Augusta, Georgia became the second town of the 13th British colony. General James Edward Oglethorpe, the colony’s founder, ordered the settlement and chose its location at the head of navigation of the Savannah River below the shoals created by the fall line. Oglethorpe’s vision was to establish an interior trading post for purchasing furs and other commodities from Native Americans to compete with New Savannah Town, a small outpost on the South Carolina side of the river.