Create baltimore.geojson
Baltimore

Novelist and Chesapeake Bay native John Dos Passos recognized the appeal of a place like Baltimore. "The character of a city as a fit place for men and women to live," he wrote, "depends on the survival of intriguing vestiges of the past." These reminders appear throughout Baltimore today: from Fort McHenry to the masterfully planned suburb of Roland Park, through the narrow streets of Fells Point, inside the 19th century Chizuk Amuno Synagogue, and on the deck of the Baltimore.