Create bor-historic-dams-water-projects.geojson
Bureau of Reclamation Historic Dams and Water Projects

The American West is known for its expansive prairies, great mountains, and arid climate. The Bureau of Reclamation, the nation’s largest supplier of water and second-largest producer of hydroelectric power, controlled the wild rivers of the West with its historic dams and water control projects--great engineering feats constructed by many workers. Damming rivers has provided water to reclaim hundreds of thousands of acres transforming sagebrush to lush farmland, and delivered water to create towns and great cities, allowing the "desert to bloom." Learn about the vital role of the Bureau of Reclamation in managing, developing, and protecting water in the United States.