Create civil-rights.geojson
Thomas Jefferson's stirring words, written in 1776 in our Declaration of Independence, defined the promise of America--freedom and equality for all. The words rang hollow, however, for the millions of African Americans held in slavery prior to the Civil War, and later denied political, economic, educational, and social equality by unjust laws and social customs. This National Register of Historic Places Travel Itinerary tells the powerful story of how and where the centuries-long struggle of African Americans to achieve the bright promise of America culminated in the mid-20th century in a heroic campaign we call the modern civil rights movement. Many of the places where these seminal events occurred, the churches, schools, homes, and neighborhoods, are listed in the National Register of Historic Places and are included in this itinerary.