Create cumberland-maryland.geojson
All Aboard for Cumberland

Known as the "Gateway to the West," Cumberland gained prominence during the 19th century as a major transportation center and as an important economic focus for the region. The city was the site of the first National Road, the western terminus of the Chesapeake & Ohio (C&O) Canal, and a center for the railroad industry. This National Register of Historic Places Travel Itinerary explores Cumberland's past through 27 historic places listed in the National Register that reflect over two centuries of history.