Create florida-shipwrecks.geojson
Florida Shipwrecks:  300 Years of Maritime History

The National Park Service's National Register of Historic Places, Archeology Program and Submerged Resources Center in partnership with the Florida Division of Historical Resources, Underwater Archaeological Section of the Bureau of Archaeological Research and the National Conference of State Historic Preservation Officers invite you to explore Florida Shipwrecks: 300 Years of Maritime History. This travel itinerary highlights 13 historic shipwrecks listed in the National Register of Historic Places that are accessible to divers and reveal the wonders and mysteries of Florida's maritime past.