Create california-coast-history.geojson
Colombus' voyages to the "New World" were just the beginning of the intermingling of peoples and cultures that formed our nation; this rich and varied history is reflected in the prehistoric and historic sites, buildings, structures, objects and districts found throughout the land. The National Register of Historic Places can guide you through our history with Discover Our Shared Heritage--a series of travel itineraries that explore our country's past through visiting places listed in the National Register of Historic Places which reflect major aspects of American history, such as exploration and settlement and cultural diversity.