Create el-camino-real-de-tierra-adentro.geojson
The history of the United States traditionally has been told from east to west. The designation of El Camino Real de Tierra Adentro as a National Historic Trail in 2000 recognizes another path, a bi-national trail, three-quarters of which winds through the central highlands of Mexico. El Camino Real tracks a different European settlement story of the U.S., one that emphasizes the shared history and heritage of Spain, Mexico and the American Southwest.