Create route-66.geojson
U.S. Highway 66 -- popularly known as Route 66 or the Mother Road -- holds a special place in American consciousness and evokes images of simpler times, mom and pop businesses, and the icons of a mobile nation on the road. This travel itinerary aids the public to visit the historic places that recall those images and experiences that are reminders of our past and evidence of the influence of the automobile. The Route 66 Discover Our Shared Heritage Travel Itinerary was produced by the National Park Service's Heritage Education Services and the National Park Service Route 66 Corridor Preservation Program, in partnership with the American Express and World Monuments Fund Sustainable Tourism Initiative and the National Conference of State Historic Preservation Officers.